Claude/debug contextbroker restart c145 g

+++ src/lib/orionld/common/orionldState.h
@@ -630,7 +630,7 @@ extern int               subCacheInterval;         // From orionld.cpp
 extern int               subCacheFlushInterval;    // From orionld.cpp
 extern bool              troe;                     // From orionld.cpp
 extern char              troeHost[256];            // From orionld.cpp
-extern unsigned short    troePort;                 // From orionld.cpp
+extern int               troePort;                 // From orionld.cpp
 extern char              troeUser[256];            // From orionld.cpp
 extern char              troePwd[256];             // From orionld.cpp
 extern char              troeSslMode[64];          // From orionld.cpp
